// panel/src/api/types.gen.ts
export type YesNo = 'yes' | 'no';
// ...
export type GetConfigsResponse = {
    [key: string]: string;
};
// ...
export type MqttAdvertiseConf = {
    MQTT_ADV_LINK_ENABLE?: YesNo;
    MQTT_ADV_LINK_BOOT?: YesNo;
    MQTT_ADV_LINK_CRON?: YesNo;
    MQTT_ADV_LINK_CRONTAB?: string;
    MQTT_ADV_LINK_TOPIC?: string;
    MQTT_ADV_LINK_RETAIN?: string;
    MQTT_ADV_LINK_QOS?: string;
    MQTT_ADV_INFO_GLOBAL_ENABLE?: YesNo;
    MQTT_ADV_INFO_GLOBAL_BOOT?: YesNo;
    MQTT_ADV_INFO_GLOBAL_CRON?: YesNo;
    MQTT_ADV_INFO_GLOBAL_CRONTAB?: string;
    MQTT_ADV_INFO_GLOBAL_TOPIC?: string;
    MQTT_ADV_INFO_GLOBAL_RETAIN?: string;
    MQTT_ADV_INFO_GLOBAL_QOS?: string;
    MQTT_ADV_CAMERA_SETTING_ENABLE?: YesNo;
    MQTT_ADV_CAMERA_SETTING_BOOT?: YesNo;
    MQTT_ADV_CAMERA_SETTING_CRON?: YesNo;
    MQTT_ADV_CAMERA_SETTING_CRONTAB?: string;
    MQTT_ADV_CAMERA_SETTING_TOPIC?: string;
    MQTT_ADV_CAMERA_SETTING_RETAIN?: string;
    MQTT_ADV_CAMERA_SETTING_QOS?: string;
    MQTT_ADV_TELEMETRY_ENABLE?: YesNo;
    MQTT_ADV_TELEMETRY_BOOT?: YesNo;
    MQTT_ADV_TELEMETRY_CRON?: YesNo;
    MQTT_ADV_TELEMETRY_CRONTAB?: string;
    MQTT_ADV_TELEMETRY_TOPIC?: string;
    MQTT_ADV_TELEMETRY_RETAIN?: string;
    MQTT_ADV_TELEMETRY_QOS?: string;
    HOMEASSISTANT_ENABLE?: YesNo;
    HOMEASSISTANT_BOOT?: YesNo;
    HOMEASSISTANT_CRON?: YesNo;
    HOMEASSISTANT_CRONTAB?: string;
    HOMEASSISTANT_MQTT_PREFIX?: string;
    HOMEASSISTANT_NAME?: string;
    HOMEASSISTANT_IDENTIFIERS?: string;
    HOMEASSISTANT_MANUFACTURER?: string;
    HOMEASSISTANT_MODEL?: string;
    HOMEASSISTANT_RETAIN?: string;
    HOMEASSISTANT_QOS?: string;
};